% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3dDigitize.main.r
\docType{package}
\name{GUImorph-package}
\alias{GUImorph-package}
\alias{GUImorph}
\title{Graphical User Interface for Morphometrics}
\description{
GUI to R programs to digitize in 3D, conduct geometric morphomteric analyses and plotting results based on OpenGL and Tk widget
}
\author{
Erik Otarola-Castillo
}
