% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3dDigitize.main.r
\docType{package}
\name{GUImorphWeb-package}
\alias{GUImorphWeb-package}
\title{Graphical User Interface for Morphometrics}
\description{
GUI to R programs to digitize in 3D, conduct geometric
  morphometric analyses and plot results, rendered in the browser with
  three.js over a local httpuv loopback server (no native OpenGL or Tk).
}
\author{
Erik Otarola-Castillo
}
